Refactor SEO keywords test suite
